+++ packages/feature-fetch/src/helper/mapper/index.ts
@@ -1,3 +1,3 @@
+export * from './map-error-to-fetch-error';
 export * from './map-error-to-network-error';
-export * from './map-error-to-service-error';
 export * from './map-response-to-request-error';