+++ packages/feature-form/src/form-field/index.ts
@@ -1,4 +1,3 @@
 export * from './create-form-field';
 export * from './create-status';
-export * from './create-validator';
 export * from './is-form-field';