+++ packages/feature-form/src/index.ts
@@ -1,7 +1,7 @@
 import type { TFormFieldStateFeature, TFormFielStatusStateFeature } from './types';
 
 export { BitwiseFlag, bitwiseFlag } from '@ibg/utils';
-export { createValidationAdapter } from 'validation-adapter';
+export { createValidator } from 'validation-adapter';
 export * from './create-form';
 export * from './form-field';
 export * from './helper';