+++ packages/openapi-router/src/__tests__/playground.test.ts
@@ -1,17 +1,34 @@
 import { Router } from 'express';
 import * as v from 'valibot';
-import { valibotAdapter } from 'validation-adapters/valibot';
+import { vValidator } from 'validation-adapters/valibot';
 import { describe, it } from 'vitest';
 
 import { createExpressOpenApiRouter } from '../features';
+import { TOpenApiRouter } from '../types';
 import { paths } from './resources/mock-openapi-types';
 
 describe('playground', () => {
 	it('types should work', async () => {
-		const openapiRouter = createExpressOpenApiRouter<paths>(Router());
+		const expressOpenApiRouter = createExpressOpenApiRouter<paths>(Router());
+		const honoOpenApiRouter: TOpenApiRouter<['base', 'hono'], paths> = null as any;
 
-		openapiRouter.get('/pet/{petId}', {
-			pathAdapter: valibotAdapter(
+		honoOpenApiRouter.get('/pet/{petId}', {
+			pathValidator: vValidator(
+				v.object({
+					petId: v.number()
+				})
+			),
+			handler: (c) => {
+				// c.req.valid('')
+				// TODO
+				// return c.json()
+
+				return c.json({ name: '', photoUrls: [] });
+			}
+		});
+
+		expressOpenApiRouter.get('/pet/{petId}', {
+			pathValidator: vValidator(
 				v.object({
 					petId: v.number()
 				})
@@ -21,8 +38,8 @@ describe('playground', () => {
 			}
 		});
 
-		openapiRouter.get('/pet/findByTags', {
-			queryAdapter: valibotAdapter(
+		expressOpenApiRouter.get('/pet/findByTags', {
+			queryValidator: vValidator(
 				v.object({
 					tags: v.optional(v.array(v.string()))
 				})