+++ packages/openapi-router/src/exceptions/ValidationError.ts
@@ -1,8 +1,8 @@
 import { type TValidationError } from 'validation-adapter';
 
-import { ServiceError } from './ServiceError';
+import { AppError } from './AppError';
 
-export class ValidationError extends ServiceError {
+export class ValidationError extends AppError {
 	constructor(errors: TValidationError[]) {
 		super('#ERR_BAD_REQUEST', 400, {
 			description: 'One or more validation errors occurred!',