+++ packages/openapi-router/src/features/with-express/with-express.ts
@@ -1,16 +1,16 @@
 import type * as express from 'express';
-import { createValidateContext, type TValidationError } from 'validation-adapter';
+import { createValidationContext, type TValidationError } from 'validation-adapter';
 import { type TOperationPathParams, type TOperationQueryParams } from '@ibg/types/openapi';
 
 import { ValidationError } from '../../exceptions';
+import { formatPath, parseParams } from '../../helper';
 import {
 	type TEnforceFeatures,
 	type TFeatureKeys,
+	type TOpenApiExpressValidators,
 	type TOpenApiRouter,
-	type TOpenApiValidationAdapters,
 	type TSelectFeatures
 } from '../../types';
-import { formatPath, parseParams } from './helper';
 
 export function withExpress<
 	GPaths extends object,
@@ -22,39 +22,35 @@ export function withExpress<
 	const expressFeatures: TSelectFeatures<['express'], GPaths> = {
 		_router: expressRouter,
 		get(this: TOpenApiRouter<['base', 'express'], GPaths>, path, config) {
-			const { handler } = config;
 			this._router.get(
-				typeof path === 'string' ? formatPath(path) : path.toString(),
+				formatPath(path),
 				parseParamsMiddleware(),
 				validationMiddleware(config),
-				requestHandler(handler as express.RequestHandler)
+				requestHandler(config.handler as express.RequestHandler)
 			);
 		},
 		post(this: TOpenApiRouter<['base', 'express'], GPaths>, path, config) {
-			const { handler } = config;
 			this._router.post(
-				typeof path === 'string' ? formatPath(path) : path.toString(),
+				formatPath(path),
 				parseParamsMiddleware(),
 				validationMiddleware(config),
-				requestHandler(handler as express.RequestHandler)
+				requestHandler(config.handler as express.RequestHandler)
 			);
 		},
 		put(this: TOpenApiRouter<['base', 'express'], GPaths>, path, config) {
-			const { handler } = config;
 			this._router.put(
-				typeof path === 'string' ? formatPath(path) : path.toString(),
+				formatPath(path),
 				parseParamsMiddleware(),
 				validationMiddleware(config),
-				requestHandler(handler as express.RequestHandler)
+				requestHandler(config.handler as express.RequestHandler)
 			);
 		},
 		del(this: TOpenApiRouter<['base', 'express'], GPaths>, path, config) {
-			const { handler } = config;
 			this._router.delete(
-				typeof path === 'string' ? formatPath(path) : path.toString(),
+				formatPath(path),
 				parseParamsMiddleware(),
 				validationMiddleware(config),
-				requestHandler(handler as express.RequestHandler)
+				requestHandler(config.handler as express.RequestHandler)
 			);
 		}
 	};
@@ -83,40 +79,40 @@ function parseParamsMiddleware(): express.RequestHandler {
 }
 
 function validationMiddleware<GPathOperation>(
-	validationAdapters: TOpenApiValidationAdapters<GPathOperation>
+	validators: TOpenApiExpressValidators<GPathOperation>
 ): express.RequestHandler {
-	const { bodyAdapter, pathAdapter, queryAdapter } = validationAdapters;
+	const { bodyValidator, pathValidator, queryValidator } = validators;
 
 	// eslint-disable-next-line @typescript-eslint/no-misused-promises -- async callback
 	return async (req, _res, next) => {
 		try {
 			const validationErrors: TValidationError[] = [];
 
-			if (bodyAdapter != null) {
-				const bodyValidationContext = createValidateContext(req.body);
-				await bodyAdapter.validate(bodyValidationContext);
+			if (bodyValidator != null) {
+				const bodyValidationContext = createValidationContext(req.body);
+				await bodyValidator.validate(bodyValidationContext);
 				for (const error of bodyValidationContext.errors) {
 					error.source = 'body';
 					validationErrors.push(error);
 				}
 			}
 
-			if (pathAdapter != null) {
-				const pathValidationContext = createValidateContext<TOperationPathParams<GPathOperation>>(
+			if (pathValidator != null) {
+				const pathValidationContext = createValidationContext<TOperationPathParams<GPathOperation>>(
 					req.params as TOperationPathParams<GPathOperation>
 				);
-				await pathAdapter.validate(pathValidationContext);
+				await pathValidator.validate(pathValidationContext);
 				for (const error of pathValidationContext.errors) {
 					error.source = 'path';
 					validationErrors.push(error);
 				}
 			}
 
-			if (queryAdapter != null) {
-				const queryValidationContext = createValidateContext<TOperationQueryParams<GPathOperation>>(
-					req.query as TOperationQueryParams<GPathOperation>
-				);
-				await queryAdapter.validate(queryValidationContext);
+			if (queryValidator != null) {
+				const queryValidationContext = createValidationContext<
+					TOperationQueryParams<GPathOperation>
+				>(req.query as TOperationQueryParams<GPathOperation>);
+				await queryValidator.validate(queryValidationContext);
 				for (const error of queryValidationContext.errors) {
 					error.source = 'query';
 					validationErrors.push(error);