+++ packages/feature-fetch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # feature-fetch
 
+## 0.0.12
+
+### Patch Changes
+
+- Renamde ValidationAdapter to Validator for more clarity
+
+## 0.0.11
+
+### Patch Changes
+
+- Improved Error class naming
+
 ## 0.0.10
 
 ### Patch Changes