+++ packages/validation-adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # validation-adapter
 
+## 0.0.5
+
+### Patch Changes
+
+- Renamde ValidationAdapter to Validator for more clarity
+
 ## 0.0.4
 
 ### Patch Changes